+++ src/sentry/incidents/serializers/alert_rule.py
@@ -47,6 +47,7 @@ from sentry.snuba.entity_subscription import (
     get_entity_subscription,
 )
 from sentry.snuba.models import QuerySubscription, SnubaQuery, SnubaQueryEventType
+from sentry.workflow_engine.migration_helpers.alert_rule import migrate_alert_rule
 
 from ...snuba.metrics.naming_layer.mri import is_mri
 from . import (
@@ -489,9 +490,10 @@ class AlertRuleSerializer(CamelSnakeModelSerializer[AlertRule]):
             )
         with transaction.atomic(router.db_for_write(AlertRule)):
             triggers = validated_data.pop("triggers")
+            user = self.context.get("user", None)
             try:
                 alert_rule = create_alert_rule(
-                    user=self.context.get("user", None),
+                    user=user,
                     organization=self.context["organization"],
                     ip_address=self.context.get("ip_address"),
                     **validated_data,
@@ -510,6 +512,12 @@ class AlertRuleSerializer(CamelSnakeModelSerializer[AlertRule]):
                 )
                 raise BadRequest
             self._handle_triggers(alert_rule, triggers)
+
+            if features.has(
+                "organizations:workflow-engine-metric-alert-processing", alert_rule.organization
+            ):
+                migrate_alert_rule(alert_rule, user)
+
             return alert_rule
 
     def update(self, instance, validated_data):
